major name refactoring; inclusion of interface for density profile
