--- single_mode/averagerays.py
@@ -1,10 +1,0 @@
-from single_mode.setup import Setup
-
-
-# 2-D evolution eqs for the average rays; single mode in the x-direction
-
-class AverageRays(Setup):
-
-    def __init__(self):
-        super().__init__()
-
