+++ single_mode/setup.py
@@ -1,9 +1,9 @@
 from helpers import constants
 from single_mode import input_parameters
-from single_mode.turbprofile import TurbProfile
+from single_mode.single_mode_turbulence_profile import SingleModeTurbulenceProfile
 
 
-class Setup(TurbProfile):
+class Setup(SingleModeTurbulenceProfile):
 
     def __init__(self):
         super().__init__(input_parameters.dne0, input_parameters.q0)
@@ -51,8 +51,8 @@ class Setup(TurbProfile):
         self.dkxdnedx = constants.dkxdnedx
         self.dkydnedx = constants.dkydnedx
 
-        self.n_profile = TurbProfile(input_parameters.dne0, input_parameters.q0)
-        self.dnedne = self.n_profile.dnedne()
+        self.turbulence_profile = SingleModeTurbulenceProfile(input_parameters.dne0, input_parameters.q0)
+        self.dnedne = self.turbulence_profile.dnedne()
         self.dnednedx = self.dnednedx()
         self.dnedxdnedx = self.dnedxdnedx()
 
@@ -60,10 +60,8 @@ class Setup(TurbProfile):
                  self.dydky0, self.dkxdkx0, self.dkxdky0, self.dkydky0, self.dxdne, self.dydne, \
                  self.dkxdne, self.dkydne, self.dxdnedx, self.dydnedx, self.dkxdnedx, self.dkydnedx
 
-
         self.num_points = input_parameters.num_points
-        self.window = input_parameters.window
-        self.abserr = input_parameters.abserr
-        self.relerr = input_parameters.relerr
+        self.abs_err = input_parameters.abs_err
+        self.rel_err = input_parameters.rel_err
         self.stop_time = input_parameters.stop_time
         self.file_name = input_parameters.file_name
